Change Bits and BitsPerSecond typedefs to double to support larger magnitudes

+++ src/ExponentialBackoff.hpp
@@ -2,7 +2,7 @@
 
 #include <random>
 
-typedef long Bits;
+typedef double Bits;
 typedef long Collisions;
 
 /**
